Import ordering and helper function to utils file
