+++ app/supplier_utils.py
@@ -30,3 +30,10 @@ def validate_agreement_details_data(agreement_details, enforce_required=True, re
 
     if errs:
         abort(400, errs)
+
+
+def check_supplier_role(role, supplier_id):
+    if role == 'supplier' and not supplier_id:
+        abort(400, "'supplierId' is required for users with 'supplier' role")
+    elif role != 'supplier' and supplier_id:
+        abort(400, "'supplierId' is only valid for users with 'supplier' role, not '{}'".format(role))
